requirements.txt
torch==2.9.0
tensorboard==2.20.0
einops==0.8.1
hydra-core==1.3.2
torchmetrics==1.8.2
torchvision==0.24.1
tqdm==4.67.1